+++ src/Ext/WeaponType/Body.cpp
@@ -127,6 +127,8 @@ void WeaponTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->Beam_Amplitude.Read(exINI, pSection, "Beam.Amplitude");
 	this->Beam_IsHouseColor.Read(exINI, pSection, "Beam.IsHouseColor");
 	this->LaserThickness.Read(exINI, pSection, "LaserThickness");
+	
+	this->CylinderRangefinding.Read(exINI, pSection, "CylinderRangefinding");
 }
 
 template <typename T>
@@ -183,6 +185,7 @@ void WeaponTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Beam_Amplitude)
 		.Process(this->Beam_IsHouseColor)
 		.Process(this->LaserThickness)
+		.Process(this->CylinderRangefinding)
 		;
 };