+++ src/Ext/WeaponType/Body.h
@@ -74,6 +74,8 @@ public:
 		Valueable<bool> Beam_IsHouseColor;
 		Valueable<int> LaserThickness;
 
+		Nullable<bool> CylinderRangefinding;
+		
 		ExtData(WeaponTypeClass* OwnerObject) : Extension<WeaponTypeClass>(OwnerObject)
 			, DiskLaser_Radius { DiskLaserClass::Radius }
 			, ProjectileRange { Leptons(100000) }
@@ -125,6 +127,7 @@ public:
 			, Beam_Amplitude { 40.0 }
 			, Beam_IsHouseColor { false }
 			, LaserThickness { 3 }
+			, CylinderRangefinding {}
 		{ }
 
 		int GetBurstDelay(int burstIndex) const;